% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_orientation}
\alias{.check_orientation}
\title{Validate orientation argument}
\usage{
.check_orientation(orientation)
}
\arguments{
\item{orientation}{Character string; must be \code{"vertical"} or
\code{"horizontal"} (case-insensitive).}
}
\value{
Normalised orientation string (lower-case).
}
\description{
Validate orientation argument
}
\keyword{internal}
