% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ecec_combine}
\alias{ecec_combine}
\title{Combine plots using an ecec_layout}
\usage{
ecec_combine(plots, layout)
}
\arguments{
\item{plots}{A list of ggplot objects.}

\item{layout}{An \code{ecec_layout} object from \code{\link{ecec_base}}.}
}
\value{
A \pkg{patchwork} composite plot.
}
\description{
Arranges a list of ggplot objects into a multi-panel figure according to
the rows and columns specified in an \code{\link{ecec_base}} layout.
}
\examples{
\dontrun{
library(ggplot2)
p1 <- ecec_scatter(mtcars, x = "wt", y = "mpg")
p2 <- ecec_scatter(mtcars, x = "hp", y = "mpg")
layout <- ecec_base(width = "2/3", height = "medium",
                    n_panels = 2, n_rows = 1, n_cols = 2)
ecec_combine(list(p1, p2), layout)
}
}
