% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ECEC_TEXT_SIZES}
\alias{ECEC_TEXT_SIZES}
\title{Default text sizes used throughout ececplots}
\usage{
ECEC_TEXT_SIZES
}
\description{
A named numeric vector that defines the default font sizes used by
\code{ecec_theme()}. All sizes are specified in points and can be scaled
globally using the \code{text_scale} argument.
}
\details{
\describe{
  \item{"axis_title"}{Axis title font size (12 pt)}
  \item{"axis_text"}{Axis tick label font size (11 pt)}
  \item{"legend_title"}{Legend title font size (12 pt)}
  \item{"legend_text"}{Legend label font size (11 pt)}
  \item{"strip_text"}{Facet strip label font size (11 pt)}
  \item{"plot_title"}{Plot title font size (14 pt)}
  \item{"plot_subtitle"}{Plot subtitle font size (12 pt)}
  \item{"plot_caption"}{Plot caption font size (11 pt)}
}
}
\keyword{internal}
