% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{print.ecec_layout}
\alias{print.ecec_layout}
\title{Print method for ecec_layout}
\usage{
\method{print}{ecec_layout}(x, ...)
}
\arguments{
\item{x}{An \code{ecec_layout} object.}

\item{...}{Unused.}
}
\description{
Print method for ecec_layout
}
